feedback1

+++ AgriYieldPredict/YieldPridict/models.py
@@ -3,5 +3,5 @@ from django.db import models
 class Feedback(models.Model):
     name = models.CharField(max_length=100)
     email = models.EmailField(max_length=100)
-    contact = models.IntegerField()
+    contact = models.BigIntegerField()
     feedback = models.CharField(max_length=250)
